Rotational Gradient
It computes the rotational gradient from 0 to 2pi of both LV and RV, using the respective axis as origin.
When the mesh is biventricular, from 0 to pi is the front of the heart. It uses the endoLV and endoRV sufraces to obtain the angles using geometry. It creates a plain with the apex and the centroids of the ventricles openings and gets the angle comparing the corresponding point to that plane.